pymd/backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio
pytest-cov==7.0.0
httpx

# Development
black==24.1.1
ruff==0.2.1
